+++ CHANGELOG.md
@@ -62,6 +62,7 @@
 - Change winesapOS log directory from "/etc/winesapos" to "/var/winesapos".
 - Change Prism Launcher to be more stable.
 - Add support for classic Snaps.
+- Add support for OneXPlayer devices.
 
 ## 4.0.0 - The Conversion Update